feat: change Float.dp2px and Float.px2dp return type

+++ module_search/src/main/kotlin/com/walisport/module/search/ui/view/FlowLayout.kt
@@ -24,8 +24,8 @@ class FlowLayout @JvmOverloads constructor(
 ) : ViewGroup(context, attrs, defStyleAttr), OnDataChangedListener {
 
     // 子項目之間的水平/垂直間距（單位為 px）
-    private val horizontalSpacing = 8f.dp2px
-    private val verticalSpacing = 8.5f.dp2px
+    private val horizontalSpacing = 8f.dp2px.toInt()
+    private val verticalSpacing = 8.5f.dp2px.toInt()
 
     // 最大可折疊行數，預設為 1 行
     private var maxFoldLines = 1
